requirements.txt
pandas
numpy
yfinance
scikit-learn
xgboost
transformers
nltk
pyYAML
logging
openpyxl
